backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
pandas==2.2.2
openpyxl==3.1.3
mysql-connector-python==8.4.0
faiss-cpu==1.8.0
openai==1.35.3
python-dotenv==1.0.1
pytesseract==0.3.10
Pillow==10.3.0
tabula-py==2.9.3
numpy==1.26.4
